requirements.txt
gitpython
prettytable
requests